chore(team): split manager archive and team tests
